refactor: move study session transition rule
