+++ src/entities/preferences/@x/study-session.ts
@@ -1,0 +1,1 @@
+export type { SwipeAction } from "../model/types";
