+++ src/entities/study-session/index.ts
@@ -1,4 +1,5 @@
 export { useStudySession, useStudySessions } from "./model/hooks";
+export { calculateNextIndex } from "./model/rules";
 export {
   clearStudySessions,
   getStudySession,
